+++ src/main/java/frc/robot/subsystems/hood/HoodIO.java
@@ -9,7 +9,7 @@ public interface HoodIO{
 
         public double positionRightActuator = 0.0;
     }
-    public default void setPosition(double position) {}
+    public default void setPosition(int position) {}
 
     // public void updateInputs(HoodIOInputsAutoLogged m_inputs);
